+++ src/agilent_vacuum/communication.py
@@ -6,7 +6,6 @@ from dataclasses import dataclass
 import logging
 import serial
 import time
-from telnetlib import Telnet
 from tenacity import *
 from threading import Thread
 from typing import Optional, Union
@@ -212,102 +211,12 @@ class SerialClient:
         self.close()
 
 
-class LanClient:
-    """
-    Lan client for communication with pump controllers.
-    Uses blocking telnet as socket.
-    """
-
-    def __init__(self, host: str, port: int = 23,
-                 timeout: float = 0.5, rate_limit: int = 10):
-        self.host = host
-        self.port = port
-        self.timeout = timeout
-        self.lock = asyncio.Lock()  # restrict to one reply request at a time
-        self.telnet = Telnet()
-        self.min_wait = 1.0 / rate_limit
-        self.last_send = time.time()
-        self.is_connected = False
-        self.reconnect_thread = Thread(target=self.reconnect, daemon=True)
-
-    def reconnect(self):
-        while self.is_connected is False:
-            time.sleep(1.0)
-            logger.info("reconnecting")
-            self.open()
-
-    def open(self):
-        self.telnet.open(host=self.host, port=self.port, timeout=self.timeout)
-        self.is_connected = True
-
-    @retry(stop=stop_after_attempt(3), wait=wait_fixed(0.2))
-    def _blocking_send(self, out_buff: bytes) -> bytes:
-        # Purge input buffer
-        try:
-            self.telnet.read_eager()
-        except EOFError:
-            pass
-
-        try:
-            self.telnet.write(out_buff)
-            reply = self.telnet.read_until(b"\x03", timeout=self.timeout)
-            checksum = self.telnet.read_eager()
-
-            if reply is None or checksum is None:
-                raise EOFError("None type response")
-
-            if len(reply) < 2 or len(checksum) < 2:
-                raise EOFError("To short response")
-
-            return reply + checksum
-
-        except BrokenPipeError as e:
-            logger.warning(f"Telnet error {e}")
-            self.is_connected = False
-            self.close()
-            time.sleep(0.1)
-            self.open()
-            logger.warning(f"Telnet tried to reconnect")
-
-    async def send(self, out_buff: bytes) -> bytes:
-        """
-        Send request and wait for reply using telnet.
-        Uses asyncio run_in_executor for blocking telnet commands
-        :param out_buff: 8 bit ascii encoded message string including STX, ETC and checksum
-        :return: response message. 8-bit ascii encoding including STX, ETX and checksum
-        """
-        if self.is_connected is False:
-            raise ComError("Client disconnected")
-
-        async with self.lock:
-            # rate limit
-            wait_time = self.min_wait - (time.time() - self.last_send)
-            if wait_time > 0.0:
-                await asyncio.sleep(wait_time)
-            # loop = asyncio.get_running_loop()
-            # in_buff = await loop.run_in_executor(None, self._blocking_send, out_buff)
-            in_buff = self._blocking_send(out_buff)
-            self.last_send = time.time()
-
-        return in_buff
-
-    def close(self) -> None:
-        """
-        Close Client
-        :return: None
-        """
-        self.telnet.close()
-
-    def __del__(self):
-        self.close()
-
-
 class AgilentDriver:
     """
     Base class for Agilent drivers
     """
 
-    def __init__(self, client: Union[LanClient, SerialClient, None], addr: int = 0, **kwargs):
+    def __init__(self, client: SerialClient | None, addr: int = 0, **kwargs):
         """
         Initialize driver
         :param addr: controller device address for RS485 communication (default 0)
